Move ShellUtils to shell package
