+++ app/src/main/java/com/termux/app/terminal/TermuxSession.java
@@ -7,7 +7,7 @@ import com.termux.app.TermuxConstants;
 import com.termux.app.TermuxService;
 import com.termux.app.utils.Logger;
 import com.termux.app.utils.PluginUtils;
-import com.termux.app.utils.ShellUtils;
+import com.termux.app.shell.ShellUtils;
 import com.termux.app.models.ExecutionCommand;
 import com.termux.terminal.TerminalSession;
